Code Review :)
it works! and its organized and not full of broken hacks. So it is good. The most noticeable missing feature is that your messages overflow, as opposed to scrolling when full. But other than that it looks good. I added a few comments if you are looking for small areas to improve.